[ modify ] ModalHeader 調整，加入功能
