+++ src/redux/boardSlice.ts
@@ -16,6 +16,7 @@ import {
 } from "@/api/boards";
 import { RootState } from "./store";
 import { newListApiAction } from "./listSlice";
+import { updateCardAction } from "./cardSlice";
 
 const initialState: {
   board: BoardsProps;
@@ -94,6 +95,17 @@ export const boardSlice = createSlice({
         state.board.list.push(newList);
       }
     });
+
+    builder.addCase(updateCardAction.fulfilled, (state, action) => {
+      const updateCard = action.payload.result;
+      state.board.list
+        .filter(({ _id }) => _id === updateCard.listId)
+        .forEach((list) => {
+          list.card = list.card.map((value) =>
+            value._id === updateCard._id ? updateCard : value
+          );
+        });
+    });
   },
 });
 
@@ -103,5 +115,10 @@ export const selectBoard = (state: RootState) => state.board.board;
 
 export const selectBoardManagers = (state: RootState) =>
   state.board.board.member.filter(({ role }) => role === "manager");
+
+export const selectListByCardId = (cardId?: string) => (state: RootState) =>
+  state.board.board.list.find(({ card }) =>
+    card.find(({ _id }) => _id === cardId)
+  );
 
 export default boardSlice.reducer;
